back up

+++ src/adm_router.js
@@ -1,27 +1,62 @@
 const express = require('express')
 var router = express.Router()
 var read_database = require('./read_database')
+const imc = require('./api/imc_api')
 const fs = require('fs')
 var Api = read_database.Api
 var Parameter = read_database.Parameter
 var Regexpr = read_database.Regexpr
 var Recommend = read_database.Recommend
+var Service = read_database.Service
 var Response = read_database.Response
 var sqlQuery = read_database.sqlQuery
 var read_DB = read_database.read_DB
 
 router.get('/', async function(req, res) {
   await read_DB()
-  return res.redirect('/adm/api')
+  if (req.session.auth && req.session.auth.usergubun == 'hoffice'){
+    return res.redirect('/adm/api')
+  }
+  return res.render('./administer/adm_login', {
+    Service
+  })
 })
 
-router.get("/view", function(req, res) {
+router.post('/login', async function(req, res) {
+  // 세션 유지 처리
+  const id = req.body.id
+  const pw = req.body.password
+  const auth = await imc.authorize({
+    id,
+    pw,
+    dancode: 1413,
+    service: 'bankdata'
+  })
+  console.log("로그인 통신 결과", auth.result[0])
+  req.session.auth = auth.result[0]
+  if (req.session.auth.usergubun == 'hoffice') {
+    return res.redirect('/adm/api')
+  } else {
+    req.session.auth = {}
+    return res.send(await alertAndRedirect(`권한이 없습니다.`, `/adm`))
+  }
+})
+
+router.get("/view", async function(req, res) {
+  if (req.session.auth.usergubun != 'hoffice'){
+    req.session.auth = {}
+    return res.send(await alertAndRedirect(`권한이 없습니다.`, `/adm`))
+  }
   return res.render("view", {
     rptfile : req.query.rptfile
   })
 })
 
 router.get('/css', async function(req, res) {
+  if (req.session.auth.usergubun != 'hoffice'){
+    req.session.auth = {}
+    return res.send(await alertAndRedirect(`권한이 없습니다.`, `/adm`))
+  }
   var css = JSON.parse(fs.readFileSync("./css/window.json"))
   try {
     return res.render("./administer/adm", {
@@ -30,11 +65,15 @@ router.get('/css', async function(req, res) {
       columns: css
     })
   } catch (e) {
-    return res.send(await alertAndRedirect('잘못된 접근 입니다.', '/adm/api'))
+    return res.send(await alertAndRedirect('잘못된 접근 입니다.', '/adm'))
   }
 })
 router.post('/css/update', async function(req, res) {
   try {
+    if (req.session.auth.usergubun != 'hoffice'){
+      req.session.auth = {}
+      return res.send(await alertAndRedirect(`권한이 없습니다.`, `/adm`))
+    }
     const width = req.body.width
     const height = req.body.height
     const window = {
@@ -46,13 +85,17 @@ router.post('/css/update', async function(req, res) {
     await fs.writeFileSync('./css/window.json', JSON.stringify(window))
     return res.send(await alertAndRedirect(`css 수정 완료`, `/adm/css`))
   } catch (e) {
-    return res.send(await alertAndRedirect('잘못된 접근 입니다.', '/adm/api'))
+    return res.send(await alertAndRedirect('잘못된 접근 입니다.', '/adm'))
   }
 })
 
 // 테이블 메인
 router.get('/:tableName', async function(req, res) {
   try {
+    if (req.session.auth.usergubun != 'hoffice'){
+      req.session.auth = {}
+      return res.send(await alertAndRedirect(`권한이 없습니다.`, `/adm`))
+    }
     const tableName = req.params.tableName
     let query = `SELECT * FROM ${tableName}`
     if (tableName == "_log")
@@ -65,13 +108,17 @@ router.get('/:tableName', async function(req, res) {
       columns
     })
   } catch (e) {
-    return res.send(await alertAndRedirect('잘못된 접근 입니다.', '/adm/api'))
+    return res.send(await alertAndRedirect('잘못된 접근 입니다.', '/adm'))
   }
 })
 
 // 등록
 router.post('/insert/:tableName', async function(req, res) {
   try {
+    if (req.session.auth.usergubun != 'hoffice'){
+      req.session.auth = {}
+      return res.send(await alertAndRedirect(`권한이 없습니다.`, `/adm`))
+    }
     const tableName = req.params.tableName
     let query = `INSERT INTO ${tableName}(`
     let idx = 0
@@ -100,6 +147,10 @@ router.post('/insert/:tableName', async function(req, res) {
 // 수정
 router.post('/update/:tableName', async function(req, res) {
   try {
+    if (req.session.auth.usergubun != 'hoffice'){
+      req.session.auth = {}
+      return res.send(await alertAndRedirect(`권한이 없습니다.`, `/adm`))
+    }
     const tableName = req.params.tableName
     let query = `UPDATE ${tableName} SET `
     let idx = 0
@@ -129,6 +180,10 @@ router.post('/update/:tableName', async function(req, res) {
 // 삭제
 router.post('/delete/:tableName', async function(req, res) {
   try {
+    if (req.session.auth.usergubun != 'hoffice'){
+      req.session.auth = {}
+      return res.send(await alertAndRedirect(`권한이 없습니다.`, `/adm`))
+    }
     const tableName = req.params.tableName
     let query = `DELETE FROM ${tableName} WHERE `
     let idx = 0
